requirements.txt
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
PyYAML==6.0.3
rich==14.2.0
yt-dlp==2025.10.22
